#14 feat: 기존 페이지 구조로 stackflow 적용
