+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@stackflow/core": "^1.0.10",
     "@stackflow/plugin-basic-ui": "^1.6.0",
+    "@stackflow/plugin-history-sync": "^1.5.0",
     "@stackflow/plugin-renderer-basic": "^1.1.8",
     "@stackflow/react": "^1.1.9",
     "@tanstack/react-query": "^5.32.1",
